Split the server module into submodules
