+++ src/server/instance.rs
@@ -1,8 +1,5 @@
-//!Server configuration and instance.
-
 use std::collections::HashMap;
 use std::net::SocketAddr;
-use std::borrow::ToOwned;
 
 use time;
 
@@ -29,178 +26,15 @@ use filter::{FilterContext, ContextFilter, ContextAction, ResponseFilter};
 use router::{Router, Endpoint};
 use handler::Handler;
 use response::Response;
-use log::{Log, StdOut};
+use log::Log;
 use header::HttpDate;
 
 use Scheme;
-use Host;
 use Global;
 use HttpResult;
+use Server;
 
 use utils;
-
-///Used to set up and run a server.
-///
-///```no_run
-///# use std::error::Error;
-///# use rustful::{Server, Handler, Context, Response};
-///# #[derive(Default)]
-///# struct R;
-///# impl Handler for R {
-///#     fn handle_request(&self, _context: Context, _response: Response) {}
-///# }
-///# let router = R;
-///let server_result = Server {
-///    host: 8080.into(),
-///    handlers: router,
-///    ..Server::default()
-///}.run();
-///
-///match server_result {
-///    Ok(_server) => {},
-///    Err(e) => println!("could not start server: {}", e.description())
-///}
-///```
-pub struct Server<R: Router> {
-    ///One or several response handlers.
-    pub handlers: R,
-
-    ///A fallback handler for when none is found in `handlers`. Leaving this
-    ///unspecified will cause an empty `404` response to be automatically sent
-    ///instead.
-    pub fallback_handler: Option<R::Handler>,
-
-    ///The host address and port where the server will listen for requests.
-    ///Default is `0.0.0.0:80`.
-    pub host: Host,
-
-    ///Use good old HTTP or the more secure HTTPS. Default is HTTP.
-    pub scheme: Scheme,
-
-    ///The number of threads to be used in the server thread pool. The default
-    ///(`None`) will cause the server to use a value based on recommendations
-    ///from the system.
-    pub threads: Option<usize>,
-
-    ///The content of the server header. Default is `"rustful"`.
-    pub server: String,
-
-    ///The default media type. Default is `text/plain, charset: UTF-8`.
-    pub content_type: Mime,
-
-    ///Tool for printing to a log. The default is to print to standard output.
-    pub log: Box<Log>,
-
-    ///Globally accessible data.
-    pub global: Global,
-
-    ///The context filter stack.
-    pub context_filters: Vec<Box<ContextFilter>>,
-
-    ///The response filter stack.
-    pub response_filters: Vec<Box<ResponseFilter>>
-}
-
-impl<R: Router> Server<R> {
-    ///Set up a new standard server. This can be useful when `handlers`
-    ///doesn't implement `Default`:
-    ///
-    ///```no_run
-    ///# use std::error::Error;
-    ///# use rustful::{Server, Handler, Context, Response};
-    ///let handler = |context: Context, response: Response| {
-    ///    //...
-    ///};
-    ///
-    ///let server_result = Server {
-    ///    host: 8080.into(),
-    ///    ..Server::new(handler)
-    ///};
-    ///```
-    pub fn new(handlers: R) -> Server<R> {
-        Server {
-            handlers: handlers,
-            fallback_handler: None,
-            host: 80.into(),
-            scheme: Scheme::Http,
-            threads: None,
-            server: "rustful".to_owned(),
-            content_type: Mime(
-                hyper::mime::TopLevel::Text,
-                hyper::mime::SubLevel::Plain,
-                vec![(hyper::mime::Attr::Charset, hyper::mime::Value::Utf8)]
-            ),
-            log: Box::new(StdOut),
-            global: Global::default(),
-            context_filters: Vec::new(),
-            response_filters: Vec::new(),
-        }
-    }
-
-    ///Start the server.
-    #[cfg(feature = "ssl")]
-    pub fn run(self) -> HttpResult<Listening> {
-        let threads = self.threads;
-        let (server, scheme) = self.build();
-        let host = server.host;
-        match scheme {
-            Scheme::Http => hyper::server::Server::http(host).and_then(|http| {
-                if let Some(threads) = threads {
-                    http.handle_threads(server, threads)
-                } else {
-                    http.handle(server)
-                }
-            }),
-            Scheme::Https {cert, key} => {
-                let ssl = try!(Openssl::with_cert_and_key(cert, key));
-                hyper::server::Server::https(host, ssl).and_then(|https| {
-                    if let Some(threads) = threads {
-                        https.handle_threads(server, threads)
-                    } else {
-                        https.handle(server)
-                    }
-                })
-            }
-        }
-    }
-
-    ///Start the server.
-    #[cfg(not(feature = "ssl"))]
-    pub fn run(self) -> HttpResult<Listening> {
-        let threads = self.threads;
-        let (server, _scheme) = self.build();
-        let host = server.host;
-        hyper::server::Server::http(host).and_then(|http| {
-            if let Some(threads) = threads {
-                http.handle_threads(server, threads)
-            } else {
-                http.handle(server)
-            }
-        })
-    }
-
-    ///Build a runnable instance of the server.
-    pub fn build(self) -> (ServerInstance<R>, Scheme) {
-        (ServerInstance {
-            handlers: self.handlers,
-            fallback_handler: self.fallback_handler,
-            host: self.host.into(),
-            server: self.server,
-            content_type: self.content_type,
-            log: self.log,
-            context_filters: self.context_filters,
-            response_filters: self.response_filters,
-            global: self.global,
-        },
-        self.scheme)
-    }
-}
-
-impl<R: Router + Default> Default for Server<R> {
-    fn default() -> Server<R> {
-        Server::new(R::default())
-    }
-}
 
 ///A runnable instance of a server.
 ///
@@ -235,10 +69,64 @@ pub struct ServerInstance<R: Router> {
     context_filters: Vec<Box<ContextFilter>>,
     response_filters: Vec<Box<ResponseFilter>>,
 
-    global: Global
+    global: Global,
 }
 
 impl<R: Router> ServerInstance<R> {
+    ///Create a new server instance, with the provided configuration. This is
+    ///the same as `Server{...}.build()`.
+    pub fn new(config: Server<R>) -> (ServerInstance<R>, Scheme) {
+        (ServerInstance {
+            handlers: config.handlers,
+            fallback_handler: config.fallback_handler,
+            host: config.host.into(),
+            server: config.server,
+            content_type: config.content_type,
+            log: config.log,
+            context_filters: config.context_filters,
+            response_filters: config.response_filters,
+            global: config.global,
+        },
+        config.scheme)
+    }
+
+    ///Start the server.
+    #[cfg(feature = "ssl")]
+    pub fn run(self, threads: Option<usize>, scheme: Scheme) -> HttpResult<Listening> {
+        let host = self.host;
+        match scheme {
+            Scheme::Http => hyper::server::Server::http(host).and_then(|http| {
+                if let Some(threads) = threads {
+                    http.handle_threads(self, threads)
+                } else {
+                    http.handle(self)
+                }
+            }),
+            Scheme::Https {cert, key} => {
+                let ssl = try!(Openssl::with_cert_and_key(cert, key));
+                hyper::server::Server::https(host, ssl).and_then(|https| {
+                    if let Some(threads) = threads {
+                        https.handle_threads(self, threads)
+                    } else {
+                        https.handle(self)
+                    }
+                })
+            }
+        }
+    }
+
+    ///Start the server.
+    #[cfg(not(feature = "ssl"))]
+    pub fn run(self, threads: Option<usize>, _scheme: Scheme) -> HttpResult<Listening> {
+        let host = self.host;
+        hyper::server::Server::http(host).and_then(|http| {
+            if let Some(threads) = threads {
+                http.handle_threads(self, threads)
+            } else {
+                http.handle(self)
+            }
+        })
+    }
 
     fn modify_context(&self, filter_storage: &mut AnyMap, context: &mut Context) -> ContextAction {
         let mut result = ContextAction::Next;
